Struct rustlearn::array::sparse::SparseArrayView

pub struct SparseArrayView<'a> {
    // some fields omitted
}

A view into a row or a column of an existing sparse matrix.

Methods

impl<'a> SparseArrayView<'a>

fn indices(&self) -> &[usize]

Returns a reference to indices of nonzero entries of the view.

fn data(&self) -> &[f32]

Returns a reference to values of nonzero entries of the view.

fn nnz(&self) -> usize

Returns the count of nonzero entries of the view.