Trait rustlearn::array::traits::RowIndex

pub trait RowIndex<Rhs> {
    type Output;
    fn get_rows(&self, index: &Rhs) -> Self::Output;
}

Trait representing a matrix whose rows can be selected to create a new matrix containing those rows.

Associated Types

type Output

Required Methods

fn get_rows(&self, index: &Rhs) -> Self::Output