Struct rustlearn::array::dense::Array

pub struct Array {
    // some fields omitted
}

Basic two-dimensional dense matrix type.

Methods

impl Array

fn zeros(rows: usize, cols: usize) -> Array

Create a rows by cols array of zeros.

fn ones(rows: usize, cols: usize) -> Array

Create a rows by cols array of ones.

fn reshape(&mut self, rows: usize, cols: usize)

Change the shape of the array to rows by cols.

Panics

If the number of elements implied by the new shape is different from the current number of elements.

fn order(&self) -> &MatrixOrder

Return the order (row-major or column-major) of the array.

fn data(&self) -> &Vec<f32>

Return an immutable reference to the underlying data buffer of the array.

The arrangement of the elements in that vector is dependent on whether this is a row-major or a column-major array.

fn as_slice(&self) -> &[f32]

Return an immutable reference to the underlying data buffer of the array.

The arrangement of the elements in that vector is dependent on whether this is a row-major or a column-major array.

fn as_mut_slice(&mut self) -> &mut [f32]

Return an mutable reference to the underlying data buffer of the array.

The arrangement of the elements in that vector is dependent on whether this is a row-major or a column-major array.

fn T(self) -> Array

Transpose the matrix.

fn sum(&self) -> f32

Compute the sum of the entries of the array.

fn mean(&self) -> f32

Compute the mean of the array.